+++ tests/test_read.py
@@ -153,7 +153,7 @@ def test_reading_En22k4_instance():  # instance from CVRPLIB
     for loc in range(1, data.num_locations):
         assert_equal(data.location(loc).service_duration, 0)
         assert_equal(data.location(loc).tw_early, 0)
-        assert_equal(data.location(loc).tw_late, np.iinfo(np.int32).max)
+        assert_equal(data.location(loc).tw_late, np.iinfo(np.int64).max)
         assert_equal(data.location(loc).release_time, 0)
         assert_equal(data.location(loc).prize, 0)
         assert_equal(data.location(loc).required, True)
